+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - Aftifact directory
+
+- Move build artifacts to dist directory under BITBUCKET_CLONE_DIR
+- Zip artifacts
+
 ## [2.1.0] - Added build arguments for node versions
 
 - MyGet feed added to npmrc